requirements.txt
langchain
langchain-community
langchain-openai
openai
scikit-learn
datasets
transformers
python-dotenv